backend/requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-extensions==3.2.1
django-ipware==4.0.2
django-test-without-migrations==0.6
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.2
python-decouple==3.6
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
simplejson==3.18.1
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
